cmd: info: Map sections to program headers
Performs a basic lookup of the start address for the section header
against the program header's virtual address and mem size to determine
which program header the section is located within.

If the section has a size of 0, it is ignored when mapping the section
header to the program header.

Signed-off-by: Kevin Townsend <kevin.townsend@analog.com>